lambda/suspend_trial_account/requirements.txt
boto3==1.24.42
GitPython==3.1.27
tenacity>=8.0.1
ruamel.yaml==0.17.21